requirements.txt
torch
agentdojo
trl[vllm]
sentence_transformers
nltk
googletrans
fast-bleu
selenium
bs4